Add php linting to the test step
Run php -l on all PHP files. This makes sure the files are at least
parsable by the current PHP version - not a comprehensive compatibility
test but a good first step.

Theoretically this could be it's own flow, but since we already have the
whole environment setup and in place and it makes no sense to run unit
tests when the files not even lint, it's simple to add it here.